rhizo_base/init.pp: Clone config from public repo
Source rccn.git from github instead of the private rhizomatica
development server. Since the repo is cloned read only, use https so
we don't have to add the github ssh host key to known_hosts. This
commit parameterizes the repo location in hiera to allow it to be
substituted with a private repo in certain deployment configurations.