rhizo_base: Use local libsofia and osmo-sip-connector
Fixes have been submitted upstream for both freeswitch's new version
of libsofia-sip-ua and osmo-sip-connector, but they are not yet
committed and released. This commit temporarily uses locally-provided
packages which allow both programs to coexist despite an otherwise
conflicting dependency on libsofia-sip-ua.

osmo-sip-connector depends on glib bindings for version 1.12.11, but
freeswitch depends on new functionality provided in 1.12.12. They used
to statically compile and keep the library in-tree, but recently moved
to a shared library architecture. Ultimately an upstream soname
version bump in the freeswitch release repo should address the
compatibility issues, but this commit provides glib bindings for
1.12.12 as well as recompling osmo-sip-connector against 1.12.12 as a
temporary stopgap.